+++ activesupport/test/cache/stores/mem_cache_store_test.rb
@@ -379,14 +379,12 @@ class MemCacheStoreTest < ActiveSupport::TestCase
     assert_equal({}, @cache.send(:read_multi_entries, [key]))
   end
 
-  def test_deprecated_connection_pool_works
-    assert_deprecated(ActiveSupport.deprecator) do
-      cache = ActiveSupport::Cache.lookup_store(:mem_cache_store, pool_size: 2, pool_timeout: 1)
-      pool = cache.instance_variable_get(:@data) # loads 'connection_pool' gem
-      assert_kind_of ::ConnectionPool, pool
-      assert_equal 2, pool.size
-      assert_equal 1, pool.instance_variable_get(:@timeout)
-    end
+  def test_pool_options_work
+    cache = ActiveSupport::Cache.lookup_store(:mem_cache_store, pool: { size: 2, timeout: 1 })
+    pool = cache.instance_variable_get(:@data) # loads 'connection_pool' gem
+    assert_kind_of ::ConnectionPool, pool
+    assert_equal 2, pool.size
+    assert_equal 1, pool.instance_variable_get(:@timeout)
   end
 
   def test_connection_pooling_by_default